Heavily restructured core, splitting it into "global data" and "frame data"
Global data is global to a core; it contains the log, task scheduler,
scene, resource loader + other facilities that are required for a core
to function.
Frame data is current to this or the past frame; modules write/modify
current data, then act/render past data - that will now be readonly - the next frame.
This simplifies the multithreaded architecture by a lot, at the expense of some extra
memory consumption.

Also modified + fixed some types in Base, Data and Event to make this possible