Major addition: stubbed out and mostly implemented the rendering backbone (OpenGL 3.3)
Most things are abstracted over the underlying API so that writing a
Vulkan backend would be hopefully easier in the future. The OpenGL
implementation is mostly complete for resource creation/deletion, but
running the actual rendering commands and FBO/pipeline passes are still not there.

Code will be documented better as the API stabilizes.