Removed EventQueue and made EventMatrix use delegates instead of queues
This change means that event servicing code must be reasonably fast not
to interrupt the program's normal operation.

Events now are basically hidden, hot-swappable functions that can be called from
any Module and trigger behaviour in any other Module that wishes to
listen to the event. They are a lower-overhead alternative to creating a Comps
that would only live one/a couple of frames (HasJustCollidedComp,
KeyJustPressedComp...) while still mantaining the flexibility to
hot-swap/remove systems altogether.

EventMatrix was moved from FrameData to GlobalData since the concept of
a "per-frame delegate" makes little sense.
Events are now as immediate as a function call instead of being shoved
into queues.

EventMatrix now contains a map that is **not** thread-safe; adding
new events from a Task could potentially create data races. The
unordered map of "event name -> event delegate" is to be replaced by a
lockless one!!